+++ R/check_compl_junctions.R
@@ -48,8 +48,8 @@ check_compl_junctions <- function(){
                     ), intern=T)
   if(any(c("prev_str03","prev_str04","prev_str05") %in% cnames)){
     message("There are complex confluences in the stream network. Please run correct_compl_junctions for correction. \n")
-    if(length(grep("prev_str",cnames)) > 3) {
-      message("There are junctions with more than three inflows. Currently, correct_compl_junctions only works for three inflows.")
+    if(length(grep("prev_str",cnames)) > 4) {
+      message("There are junctions with more than four inflows. Currently, correct_compl_junctions only works for up to four inflows.")
     }
     ret <- TRUE
   }